Renamed scan analysis to avoid naming conflicts
