+++ GEECS-PythonAPI/geecs_python_api/controls/data_acquisition/__init__.py
@@ -1,4 +1,4 @@
 from .data_acquisition import DeviceManager, DataLogger, DataInterface, ActionManager, SoundPlayer
 from .scan_manager import ScanManager
 from .utils import ConsoleLogger, get_full_config_path, visa_config_generator
-from .scan_analysis import ScanAnalysis, MagSpecStitcherAnalysis  # Import the necessary analysis classes
+from .scan_analysis_routines import ScanAnalysis, MagSpecStitcherAnalysis  # Import the necessary analysis classes
